feat: Implement unified search architecture with bug fixes
This commit introduces a comprehensive redesign of aMule's search subsystem
and resolves several critical bugs.

## New Features

### Unified Search Architecture
- Modular search framework with pluggable engines (Local, Global, Kad)
- Clean separation between controllers, packet builders, and result handlers
- Centralized UnifiedSearchManager for coordinating search operations
- Search state management with proper timeout handling
- Search result caching for improved performance

### New Components
- SearchStateManager: Manages search state transitions
- SearchTimeoutManager: Handles search timeouts and auto-retry
- SimpleSearchCache: Caches search results for reuse
- ParallelSearch: Enables parallel Kad searches
- Protocol conversion layer for ED2K/Kad interoperability

## Bug Fixes

### Search Issues
- Fix Kad search stuck at "Searching" state
- Fix search tabs showing incorrect status
- Fix "More" button errors due to missing searchType conversion
- Fix corrupted FileID in search results from truncated packets
- Fix race conditions causing search initialization failures
- Fix search type being lost during timeout handling

### Encoding Issues
- Fix double-encoded UTF-8 in network data
- Fix corrupted text from partial UTF-8 conversion failures
- Remove Unicode replacement characters (U+FFFD) from log output
- Fix format specifier mismatches and invalid Unicode crashes
- Fix format string crashes in SearchList and codec logging

### Security & Stability
- Add comprehensive input validation for network packets
- Fix critical UDP socket crash in Boost ASIO implementation
- Fix critical security vulnerabilities and race conditions

## Testing
- Add unit tests for unified search core abstractions
- Add ModernLoggingTest for logging infrastructure

## Documentation
- Add comprehensive search architecture documentation
- Add build guide and implementation summaries

+++ src/ClientList.cpp
@@ -1008,30 +1008,30 @@ void CClientList::CleanUpClientList()
 				if (!(pCurClient->GetUploadState() == US_NONE || (pCurClient->GetUploadState() == US_BANNED && !pCurClient->IsBanned()))) {
 					AddDebugLogLineN(logProxy,
 						CFormat(wxT("Debug: Not deleted client %x with up state: %i "))
-							% (long int)pCurClient % pCurClient->GetUploadState());
+							% (size_t)pCurClient % pCurClient->GetUploadState());
 				}
 				if (!(pCurClient->GetDownloadState() == DS_NONE)) {
 					AddDebugLogLineN(logProxy,
 						CFormat(wxT("Debug: Not deleted client %x with down state: %i "))
-							% (long int)pCurClient % pCurClient->GetDownloadState());
+							% (size_t)pCurClient % pCurClient->GetDownloadState());
 				}
 				if (!(pCurClient->GetChatState() == MS_NONE)) {
 					AddDebugLogLineN(logProxy,
 						CFormat(wxT("Debug: Not deleted client %x with chat state: %i "))
-							% (long int)pCurClient % pCurClient->GetChatState());
+							% (size_t)pCurClient % pCurClient->GetChatState());
 				}
 				if (!(pCurClient->GetKadState() == KS_NONE)) {
 					AddDebugLogLineN(logProxy,
 						CFormat(wxT("Debug: Not deleted client %x with kad state: %i ip: %s"))
-							% (long int)pCurClient % (int)pCurClient->GetKadState() % pCurClient->GetFullIP());
+							% (size_t)pCurClient % (int)pCurClient->GetKadState() % pCurClient->GetFullIP());
 				}
 				if (!(pCurClient->GetSocket() == NULL)) {
 					AddDebugLogLineN(logProxy,
-						CFormat(wxT("Debug: Not deleted client %x: has socket")) % (long int)pCurClient);
+						CFormat(wxT("Debug: Not deleted client %x: has socket")) % (size_t)pCurClient);
 				}
 				AddDebugLogLineN(logProxy,
 					CFormat(wxT("Debug: Not deleted client %x with kad version: %i"))
-						% (long int)pCurClient % pCurClient->GetKadVersion());
+						% (size_t)pCurClient % pCurClient->GetKadVersion());
 #endif
 			}
 		}
